webpack improvements * adds __DEV__; fixes to package.json; * webpack.config.babel
