--- webpack.config.js
@@ -1,3 +1,0 @@
-require('babel-register');
-
-module.exports = require('./client/webpack.config');
